backend/requirements.txt
# TrustChain Backend Dependencies
# Python 3.11+ required

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI Provider SDKs
anthropic==0.18.1
openai==1.10.0

# HTTP client for Ollama
aiohttp==3.9.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Async utilities
asyncio==3.4.3

# Logging and monitoring
structlog==24.1.0

# Testing (development)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code quality (development)
black==24.1.1
ruff==0.1.14
mypy==1.8.0
